support for multiple Arnold property modifiers
Originally the code had a break once it found the first Arnold property modifier, removed that in case you have multiple Arnold property modifiers. Not sure why you'd want to have multiple ones. From a quick test it seems that it uses the values from the last one in the modifier stack.